Added a locate command to help locate honeypot blocks. This is not tested yet as I created this on my MacBook and not my dev PC so I'm not sure if it functions yet and won't know until I can test it in a couple days. For now, it theoretically should function, but with some potentially odd results.
IT IS ALSO NOT OPTIMIZED IN THE SLIGHTEST. It will manually look up each and every block within 5x5x5 cube of the player executing the command. That is a total of 125 total potential "database" lookups, and it's just doing string comparison. It's not great. The only way to truly improve this is to improve the way I store data, which would require a re-write of the storage manager classes and also figure out how to upgrade already-existing databases. It shouldn't be too difficult to perform the upgrade, but rewriting the storage classes could prove to be difficult. Who knows, it's a challenge :)