Renamed some files to move away from the "Honeypot..." naming convention as it's hard to read and annoying to type. Started working on implementing SQLite support by splitting the JSON storage code into its own class and adding a config value for it.
Unfortunately, the way I've written the code means I have to re-write the entire storage manager classes due to them being heavily reliant on ArrayLists which are extremely volatile and aren't necessary for use in SQLite. Luckily, re-writing the Storage Manager classes doesn't mean much work in the rest of the plugin since these classes were specifically written so I don't have to do any CRUD actions within other class. As long as I maintain the same methods nothing should break. I just need to re-write the methods to support SQLite.

Currently, these methods ONLY search within the ArrayLists and use the JSON file to permanently store items. The JSON is only loaded once into an ArrayList, and that ArrayList is modified multiple times. Each modification also saves to the JSON file, but nothing is ever actually pulled from JSON after the original load. This means it's extremely fast, but impossible to extend to other types of storage types unless I sacrifice functionality of those types (Such as SQLite, where the whole point of using it is for SELECT statements) and can corrupt if the server runs out of memory or if power is pulled unexpectedly from the server.