Honeypot 1.6 Gotta love when I get in the programming mood lol! One day after version 1.5 comes version 1.6! This version _finally_ adds custom actions to Honeypot! No more default, built in actions. You can finally teleport a honeypot breaker to the void, or spawn 1000 creepers around them, or broadcast a message to all players! The possibilities are endless!
Note: This feature is **DISABLED** by default. The reason being is that actions are run as the server itself, not as any specific staff member or player. This means you can op, deop, stop, restart, or do whatever you want to the server with these blocks. Please exercise caution when using this feature. As it can be dangerous, it also requires a specific permission to use, honeypot.custom. Without this permission, staff will be able to create all types of honeypots EXCEPT custom action ones. I am working on a solution to this, but it is difficult without severely hurting the functionality of this feature.

Added:
* Custom actions to Honeypots! Now you are able to do whatever you want with these blocks. This feature uses the Conversation API which is, unfortunately, poorly documented, BUT I still got it to work! :)
* Use %player% in custom honeypot block actions to target the player that activated the custom action.

Changed:
* Removed (more) unnecessary parts of code
* /honeypot create custom now requires players to have a specific "honeypot.custom" permission to use. Otherwise, staff can create honeypot blocks except for custom action ones.
* Incremented config version to '2'.

Bug Fixes:
* None in this version! I couldn't find any bugs, but if you find any please report them!

Known Issues:
* When the config is updated upon plugin load, new values are put at the bottom. This is unfortunate because I made it so pretty last time D: I am working to see if there is a way to resolve this in the updater settings. For now, though, the config will update fine but newer options may not be where you expect them to be.